// openapi/collector/model/twitterTweetQueryRequest.ts
/**
 * 
 * MPDC Collector API
 *
 * The version of the OpenAPI document: 0.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TwitterTweetQueryRequest { 
    apiKey: string;
    fetchMultimedia: Array<TwitterTweetQueryRequest.FetchMultimediaEnum>;
    indexInCineast: boolean;
    label: string;
    collectionName: string;
    from: string;
    to: string;
    keywords?: Array<string>;
    accounts?: Array<string>;
    queryDelayMinutes?: number;
    referencedTweetsDepth?: number;
    useStreamingApiIfPossible?: boolean;
}
export namespace TwitterTweetQueryRequest {
    export type FetchMultimediaEnum = 'UNKNOWN' | 'ANIMATED_IMAGE' | 'IMAGE' | 'VIDEO';
    export const FetchMultimediaEnum = {
        Unknown: 'UNKNOWN' as FetchMultimediaEnum,
        AnimatedImage: 'ANIMATED_IMAGE' as FetchMultimediaEnum,
        Image: 'IMAGE' as FetchMultimediaEnum,
        Video: 'VIDEO' as FetchMultimediaEnum
    };
}